fix(tests): register --write-outputs pytest option globally (#251)
Move pytest_addoption() and pytest_configure() hooks from the deeply-nested
engine/conftest.py to a root-level conftest.py. Pytest discovers conftest.py
via the rootdir and testpaths configuration during plugin loading, before
argument parsing. The engine conftest was never being loaded because:

1. No root-level conftest.py exists
2. No testpaths configuration guides pytest to engine/
3. Pytest doesn't walk arbitrary src/ subdirectories

By placing the hook implementations at the project root, pytest discovers them
early and registers the --write-outputs option before argparse runs, allowing
the flag to be recognized.
