+++ osprey_worker/src/osprey/engine/conftest.py
@@ -407,35 +407,3 @@ def check_failure(check_output: CheckOutputFunction) -> CheckFailureFunction:
     return check_failure
 
 
-def pytest_addoption(parser: 'Parser') -> None:
-    parser.addoption(
-        '--write-outputs', action='store_true', help='write checked validator outputs instead of checking them'
-    )
-
-
-def pytest_configure(config: 'Config') -> None:
-    config.addinivalue_line(
-        'markers',
-        'use_validators([validator_classes, ...]): used with the `run_validation` fixture, '
-        'to specify the validators to execute.',
-    )
-    config.addinivalue_line(
-        'markers',
-        'use_standard_rules_validators(): used with the `run_validation` fixture, runs all normal rules validators.',
-    )
-    config.addinivalue_line(
-        'markers',
-        'inject_validator_result(validator={...}, result={...}): used with the `run_validation` fixture, '
-        'to mock requisite validator results.',
-    )
-    config.addinivalue_line(
-        'markers', 'use_udf_registry(UDFRegistry()): use a given udf registry during validation/execution'
-    )
-    config.addinivalue_line(
-        'markers', 'use_osprey_stdlib(): uses the osprey udf stdlib registry during validation/execution.'
-    )
-    config.addinivalue_line(
-        'markers',
-        'vary_output_by_py_version(): instructs the `check_output` feature to record a separate file for '
-        'different python major/minor versions',
-    )
